requirements.txt
# Core Testing Framework
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Playwright
playwright==1.40.0
pytest-playwright==0.4.3

# API Testing
requests==2.31.0
httpx==0.25.2
jsonschema==4.20.0

# Reporting
allure-pytest==2.13.2
pytest-reportportal==5.3.2

# Logging & Utilities
loguru==0.7.2
python-dotenv==1.0.0
faker==20.1.0
pandas==2.1.3

# AI Integration
openai
langchain
langchain-openai

# Data & Config
pydantic==2.5.0
pyyaml==6.0.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0